refactor(Workspace): Move workspace actions
Move workspace actions to actions module
