+++ MiAZ/data/docs/VERSION
@@ -1,1 +1,1 @@
-0.0.1+build.5933
+0.0.1+build.5934
