+++ MiAZ/frontend/desktop/actions.py
@@ -3,7 +3,9 @@
 
 import os
 import glob
+from datetime import datetime
 
+from gi.repository import GLib
 from gi.repository import GObject
 from gi.repository import Gtk
 
@@ -90,64 +92,136 @@ class MiAZActions(GObject.GObject):
         self.app.show_stack_page_by_name('rename')
 
     def document_rename_multiple(self, item_type, items):
+
+        def update_columnview(dropdown, gparamobj, columnview, item_type, items):
+            i_type = item_type.__gtype_name__
+            i_title = item_type.__title__
+            citems = []
+            for item in items:
+                try:
+                    source = item.id
+                    name, ext = self.util.filename_details(source)
+                    n = Field[item_type]
+                    tmpfile = name.split('-')
+                    tmpfile[n] = dropdown.get_selected_item().id
+                    filename = "%s.%s" % ('-'.join(tmpfile), ext)
+                    target = os.path.join(os.path.dirname(source), filename)
+                    txtId = "<small>%s</small>" % os.path.basename(source)
+                    txtTitle = "<small>%s</small>" % os.path.basename(target)
+                    citems.append(File(id=txtId, title=txtTitle))
+                except Exception as error:
+                    # FIXME: AtributeError: 'NoneType' object has no attribute 'id'
+                    # It happens when managing resources from inside the dialog
+                    pass
+            columnview.update(citems)
+
         def update_dropdown(config, dropdown, item_type, any_value):
             title = item_type.__gtype_name__
             self.dropdown_populate(dropdown, item_type, any_value=any_value)
             dropdown.set_selected(0)
 
+        def calendar_day_selected(calendar, label, columnview, items):
+            adate = calendar.get_date()
+            y = "%04d" % adate.get_year()
+            m = "%02d" % adate.get_month()
+            d = "%02d" % adate.get_day_of_month()
+            sdate = "%s%s%s" % (y, m, d)
+            ddate = datetime.strptime(sdate, '%Y%m%d')
+            label.set_text(ddate.strftime('%A, %B %d %Y'))
+            citems = []
+            for item in items:
+                source = os.path.basename(item.id)
+                name, ext = self.util.filename_details(source)
+                lname = name.split('-')
+                lname[0] = sdate
+                target = "%s.%s" % ('-'.join(lname), ext)
+                citems.append(File(id=source, title=target))
+            columnview.update(citems)
+
+        def dialog_response(dialog, response, dropdown, item_type, items):
+            if response == Gtk.ResponseType.ACCEPT:
+                title = item_type.__title__
+                for item in items:
+                    source = item.id
+                    name, ext = self.util.filename_details(source)
+                    n = Field[item_type]
+                    tmpfile = name.split('-')
+                    tmpfile[n] = dropdown.get_selected_item().id
+                    target = "%s.%s" % ('-'.join(tmpfile), ext)
+                    self.util.filename_rename(source, target)
+            dialog.destroy()
+
+        def dialog_response_date(dialog, response, calendar, items):
+            if response == Gtk.ResponseType.ACCEPT:
+                adate = calendar.get_date()
+                y = "%04d" % adate.get_year()
+                m = "%02d" % adate.get_month()
+                d = "%02d" % adate.get_day_of_month()
+                sdate = "%s%s%s" % (y, m, d)
+                for item in items:
+                    source = os.path.basename(item.id)
+                    name, ext = self.util.filename_details(source)
+                    lname = name.split('-')
+                    lname[0] = sdate
+                    target = "%s.%s" % ('-'.join(lname), ext)
+                    self.util.filename_rename(source, target)
+            dialog.destroy()
+
         self.factory = self.app.get_factory()
         self.config = self.backend.conf
 
-        i_type = item_type.__gtype_name__
-        i_title = item_type.__title__
-        self.log.debug("Rename %s for:", i_title)
-        box = self.factory.create_box_vertical(spacing=6, vexpand=True, hexpand=True)
-        label = self.factory.create_label('Rename %d files by setting the field <b>%s</b> to:\n' % (len(items), i_title))
-        dropdown = self.factory.create_dropdown_generic(item_type)
-        btnManage = self.factory.create_button('miaz-res-manage', '')
-        btnManage.connect('clicked', self.on_resource_manage, Configview[i_type](self.app))
-        frame = Gtk.Frame()
-        cv = MiAZColumnViewMassRename(self.app)
-        cv.get_style_context().add_class(class_name='monospace')
-        cv.set_hexpand(True)
-        cv.set_vexpand(True)
-        dropdown.connect("notify::selected-item", self.update_columnview, cv, item_type, items)
-        self.config[i_type].connect('used-updated', update_dropdown, dropdown, item_type, False)
-        self.dropdown_populate(dropdown, item_type, any_value=False)
-        frame.set_child(cv)
-        box.append(label)
-        hbox = self.factory.create_box_horizontal()
-        hbox.append(dropdown)
-        hbox.append(btnManage)
-        box.append(hbox)
-        box.append(frame)
-        dialog = self.factory.create_dialog_question(self.app.win, 'Mass renaming', box, width=1024, height=600)
-        # ~ dialog.connect('response', self._on_mass_action_rename_response, dropdown, item_type)
-        dialog.show()
-
-    def update_columnview(self, dropdown, gparamobj, columnview, item_type, items):
-        i_type = item_type.__gtype_name__
-        i_title = item_type.__title__
-        citems = []
-        for item in items:
-            try:
-                source = item.id
-                name, ext = self.util.filename_details(source)
-                n = Field[item_type]
-                tmpfile = name.split('-')
-                tmpfile[n] = dropdown.get_selected_item().id
-                filename = "%s.%s" % ('-'.join(tmpfile), ext)
-                target = os.path.join(os.path.dirname(source), filename)
-                txtId = "<small>%s</small>" % os.path.basename(source)
-                txtTitle = "<small>%s</small>" % os.path.basename(target)
-                citems.append(File(id=txtId, title=txtTitle))
-            except Exception as error:
-                # FIXME: AtributeError: 'NoneType' object has no attribute 'id'
-                # It happens when managing resources from inside the dialog
-                self.log.error(error)
-        columnview.update(citems)
-
-    # ~ self.log.debug("%s > %s", item_type, items)
+        if item_type != Date:
+            i_type = item_type.__gtype_name__
+            i_title = item_type.__title__
+            self.log.debug("Rename %s for:", i_title)
+            box = self.factory.create_box_vertical(spacing=6, vexpand=True, hexpand=True)
+            label = self.factory.create_label('Rename %d files by setting the field <b>%s</b> to:\n' % (len(items), i_title))
+            dropdown = self.factory.create_dropdown_generic(item_type)
+            btnManage = self.factory.create_button('miaz-res-manage', '')
+            btnManage.connect('clicked', self.on_resource_manage, Configview[i_type](self.app))
+            frame = Gtk.Frame()
+            cv = MiAZColumnViewMassRename(self.app)
+            cv.get_style_context().add_class(class_name='monospace')
+            cv.set_hexpand(True)
+            cv.set_vexpand(True)
+            dropdown.connect("notify::selected-item", update_columnview, cv, item_type, items)
+            self.config[i_type].connect('used-updated', update_dropdown, dropdown, item_type, False)
+            self.dropdown_populate(dropdown, item_type, any_value=False)
+            frame.set_child(cv)
+            box.append(label)
+            hbox = self.factory.create_box_horizontal()
+            hbox.append(dropdown)
+            hbox.append(btnManage)
+            box.append(hbox)
+            box.append(frame)
+            dialog = self.factory.create_dialog_question(self.app.win, 'Mass renaming', box, width=1024, height=600)
+            dialog.connect('response', dialog_response, dropdown, item_type, items)
+            dialog.show()
+        else:
+            self.log.debug("FIXME! Not implemented")
+            box = self.factory.create_box_vertical(spacing=6, vexpand=True, hexpand=True)
+            hbox = self.factory.create_box_horizontal()
+            label = Gtk.Label()
+            calendar = Gtk.Calendar()
+            btnDate = self.factory.create_button_popover(icon_name='miaz-res-date', widgets=[calendar])
+            hbox.append(btnDate)
+            hbox.append(label)
+            frame = Gtk.Frame()
+            cv = MiAZColumnViewMassRename(self.app)
+            cv.get_style_context().add_class(class_name='monospace')
+            cv.set_hexpand(True)
+            cv.set_vexpand(True)
+            frame.set_child(cv)
+            box.append(hbox)
+            box.append(frame)
+            sdate = datetime.strftime(datetime.now(), '%Y%m%d')
+            iso8601 = "%sT00:00:00Z" % sdate
+            calendar.connect('day-selected', calendar_day_selected, label, cv, items)
+            calendar.select_day(GLib.DateTime.new_from_iso8601(iso8601))
+            calendar.emit('day-selected')
+            dialog = self.factory.create_dialog_question(self.app.win, 'Mass renaming', box, width=640, height=480)
+            dialog.connect('response', dialog_response_date, calendar, items)
+            dialog.show()
 
     def select_dropdown_item(self, dropdown, key):
         found = False
