+++ MiAZ/frontend/desktop/widgets/workspace.py
@@ -120,97 +120,6 @@ class MiAZWorkspace(Gtk.Box):
         title = item_type.__gtype_name__
         self.actions.dropdown_populate(self.dropdown[title], item_type)
 
-    def _on_mass_action_rename_dialog_date(self, *args):
-        def calendar_day_selected(calendar, label, columnview, items):
-            adate = calendar.get_date()
-            y = "%04d" % adate.get_year()
-            m = "%02d" % adate.get_month()
-            d = "%02d" % adate.get_day_of_month()
-            sdate = "%s%s%s" % (y, m, d)
-            ddate = datetime.strptime(sdate, '%Y%m%d')
-            label.set_text(ddate.strftime('%A, %B %d %Y'))
-            citems = []
-            for item in items:
-                source = os.path.basename(item.id)
-                name, ext = self.util.filename_details(source)
-                lname = name.split('-')
-                lname[0] = sdate
-                target = "%s.%s" % ('-'.join(lname), ext)
-                citems.append(File(id=source, title=target))
-            columnview.update(citems)
-        box = self.factory.create_box_vertical(spacing=6, vexpand=True, hexpand=True)
-        hbox = self.factory.create_box_horizontal()
-        label = Gtk.Label()
-        calendar = Gtk.Calendar()
-        btnDate = self.factory.create_button_popover(icon_name='miaz-res-date', widgets=[calendar])
-        hbox.append(btnDate)
-        hbox.append(label)
-        frame = Gtk.Frame()
-        cv = MiAZColumnViewMassRename(self.app)
-        cv.get_style_context().add_class(class_name='monospace')
-        cv.set_hexpand(True)
-        cv.set_vexpand(True)
-        frame.set_child(cv)
-        box.append(hbox)
-        box.append(frame)
-        sdate = datetime.strftime(datetime.now(), '%Y%m%d')
-        iso8601 = "%sT00:00:00Z" % sdate
-        calendar.connect('day-selected', calendar_day_selected, label, cv, self.selected_items)
-        calendar.select_day(GLib.DateTime.new_from_iso8601(iso8601))
-        calendar.emit('day-selected')
-        dialog = self.factory.create_dialog_question(self.app.win, 'Mass renaming', box, width=640, height=480)
-        dialog.connect('response', self._on_mass_action_rename_date_response, calendar)
-        dialog.show()
-
-    def _on_mass_action_rename_date_response(self, dialog, response, calendar):
-        if response == Gtk.ResponseType.ACCEPT:
-            adate = calendar.get_date()
-            y = "%04d" % adate.get_year()
-            m = "%02d" % adate.get_month()
-            d = "%02d" % adate.get_day_of_month()
-            sdate = "%s%s%s" % (y, m, d)
-            for item in self.selected_items:
-                source = os.path.basename(item.id)
-                name, ext = self.util.filename_details(source)
-                lname = name.split('-')
-                lname[0] = sdate
-                target = "%s.%s" % ('-'.join(lname), ext)
-                self.util.filename_rename(source, target)
-        dialog.destroy()
-
-    # ~ def _on_mass_action_rename_dialog(self, action, data, item_type):
-        # ~ def update_dropdown(config, dropdown, item_type, any_value):
-            # ~ title = item_type.__gtype_name__
-            # ~ self.actions.dropdown_populate(dropdown, item_type, any_value=any_value)
-            # ~ dropdown.set_selected(0)
-
-        # ~ i_type = item_type.__gtype_name__
-        # ~ i_title = item_type.__title__
-        # ~ self.log.debug("Rename %s for:", i_title)
-        # ~ box = self.factory.create_box_vertical(spacing=6, vexpand=True, hexpand=True)
-        # ~ label = self.factory.create_label('Rename %d files by setting the field <b>%s</b> to:\n' % (len(self.selected_items), i_title))
-        # ~ dropdown = self.factory.create_dropdown_generic(item_type)
-        # ~ btnManage = self.factory.create_button('miaz-res-manage', '')
-        # ~ btnManage.connect('clicked', self.on_resource_manage, Configview[i_type](self.app))
-        # ~ frame = Gtk.Frame()
-        # ~ cv = MiAZColumnViewMassRename(self.app)
-        # ~ cv.get_style_context().add_class(class_name='monospace')
-        # ~ cv.set_hexpand(True)
-        # ~ cv.set_vexpand(True)
-        # ~ dropdown.connect("notify::selected-item", self._on_mass_action_rename_update_dropdown, cv, item_type)
-        # ~ self.config[i_type].connect('%s-used' % i_type.lower(), update_dropdown, dropdown, item_type, False)
-        # ~ self.actions.dropdown_populate(dropdown, item_type, any_value=False)
-        # ~ frame.set_child(cv)
-        # ~ box.append(label)
-        # ~ hbox = self.factory.create_box_horizontal()
-        # ~ hbox.append(dropdown)
-        # ~ hbox.append(btnManage)
-        # ~ box.append(hbox)
-        # ~ box.append(frame)
-        # ~ dialog = self.factory.create_dialog_question(self.app.win, 'Mass renaming', box, width=1024, height=600)
-        # ~ dialog.connect('response', self._on_mass_action_rename_response, dropdown, item_type)
-        # ~ dialog.show()
-
     def on_resource_manage(self, widget: Gtk.Widget, selector: Gtk.Widget):
         box = self.factory.create_box_vertical(spacing=0, vexpand=True, hexpand=True)
         box.append(selector)
@@ -219,28 +128,6 @@ class MiAZWorkspace(Gtk.Box):
         selector.update()
         dialog = self.factory.create_dialog(self.app.win, 'Manage %s' % config_for, box, 800, 600)
         dialog.show()
-
-    # ~ def _on_mass_action_rename_update_dropdown(self, dropdown, item, columnview, item_type):
-        # ~ i_type = item_type.__gtype_name__
-        # ~ i_title = item_type.__title__
-        # ~ citems = []
-        # ~ for item in self.selected_items:
-            # ~ try:
-                # ~ source = item.id
-                # ~ name, ext = self.util.filename_details(source)
-                # ~ n = Field[item_type]
-                # ~ tmpfile = name.split('-')
-                # ~ tmpfile[n] = dropdown.get_selected_item().id
-                # ~ filename = "%s.%s" % ('-'.join(tmpfile), ext)
-                # ~ target = os.path.join(os.path.dirname(source), filename)
-                # ~ txtId = "<small>%s</small>" % os.path.basename(source)
-                # ~ txtTitle = "<small>%s</small>" % os.path.basename(target)
-                # ~ citems.append(File(id=txtId, title=txtTitle))
-            # ~ except AttributeError:
-                # ~ # FIXME: AtributeError: 'NoneType' object has no attribute 'id'
-                # ~ # It happens when managing resources from inside the dialog
-                # ~ pass
-        # ~ columnview.update(citems)
 
     def _on_mass_action_project_dialog(self, *args):
         def update_dropdown(config, dropdown, item_type, any_value):
@@ -312,18 +199,7 @@ class MiAZWorkspace(Gtk.Box):
             self.projects.add_batch(pid, docs)
         dialog.destroy()
 
-    def _on_mass_action_rename_response(self, dialog, response, dropdown, item_type):
-        if response == Gtk.ResponseType.ACCEPT:
-            title = item_type.__title__
-            for item in self.selected_items:
-                source = item.id
-                name, ext = self.util.filename_details(source)
-                n = Field[item_type]
-                tmpfile = name.split('-')
-                tmpfile[n] = dropdown.get_selected_item().id
-                target = "%s.%s" % ('-'.join(tmpfile), ext)
-                self.util.filename_rename(source, target)
-        dialog.destroy()
+
 
     def _on_filters_toggled(self, button, data=None):
         active = button.get_active()
@@ -761,6 +637,3 @@ class MiAZWorkspace(Gtk.Box):
         self.view.column_subtitle.set_expand(True)
         self.view.refilter()
         self.tgbFilters.set_visible(True)
-
-
-
